+++ src/ray/castCone.js
@@ -9,7 +9,7 @@
  * @param {object} [options] - options that may include:
  * @param {object[]} [options.objects = Raycaster.mappedObjects] - Array of game objects to test. If not provided test all mapped game objects.
  *
- * @return {Phaser.Geom.Point[]} Array of points of ray's closest intersections with tested objects. Additionally each point contains reference to hit mapped object and it's segment if available.
+ * @return {Phaser.Math.Vector2[]} Array of points of ray's closest intersections with tested objects. Additionally each point contains reference to hit mapped object and it's segment if available.
  */
 export function castCone(options = {}) {
     let originalAngle = this.angle;
@@ -141,7 +141,7 @@ export function castCone(options = {}) {
 
                     if(Math.abs(angleOffsetDeg) < Phaser.Math.RadToDeg(cone / 2)) {
                         rayTargets.push({
-                            point: new Phaser.Geom.Point(intersection.x, intersection.y),
+                            point: new Phaser.Math.Vector2(intersection.x, intersection.y),
                             angle: Phaser.Math.Angle.Between(this.origin.x, this.origin.y, intersection.x, intersection.y),
                             angleOffsetDeg: -angleOffsetDeg
                         });
@@ -188,11 +188,11 @@ export function castCone(options = {}) {
             //if intersection hits target point check if ray "glanced" mapped object.
             let castSides = false;
             if(this.round) {
-                let roundedTarget = new Phaser.Geom.Point(Math.round(target.point.x), Math.round(target.point.y));
-                castSides = Phaser.Geom.Point.Equals(roundedTarget, intersection)
+                let roundedTarget = new Phaser.Math.Vector2(Math.round(target.point.x), Math.round(target.point.y));
+                castSides = roundedTarget.equals(intersection)
             }
             else {
-                castSides = Phaser.Geom.Point.Equals(target.point, intersection);
+                castSides = target.point.equals(intersection);
             }
             
             if(!castSides) {